feat!: setToken renamed to setCurrentPage
setCurrentPage accepts a PageToken, a PageOffset or a PageNumber object.
PageToken always require at least two order rules, PageOffset and
PageNumber does not require any order rule (it is, however, recommended).

Performance considerations:
When executing against a database, PageOffset and PageNumber issue a
query with limit and offset set; setting an offset decreses the query
performance and is very inefficient compared to a where-based pagination.